+++ bin/cli.js
@@ -15,92 +15,15 @@ const styles = {
 }
 
 program.version(manifest.version)
-  .arguments('<filePath> <componentName>')
-  .action(function(filePath, componentName){
-    try {
-      let normalizePath = path.normalize(filePath);
-      fs.readFile(normalizePath, "utf-8", function (err, data) {
-        if (!data) {
-          console.error('can\'t resolve filePath: ' + normalizePath);
-          return;
-        }
-        let ast = astHelper.flowAst(data);
-        let params = {
-          name: componentName
-        };
-        // merge config to options
-        let options = Object.assign({}, readConfig(), params);
-        Promise.all([
-          actions.findComponentNode(ast, options),
-          actions.findPropTypesNode(ast, options),
-          actions.findPropTypesNode(ast, Object.assign({}, options, {
-            alias: 'defaultProps'
-          }))
-        ], options).then((nodes) => {
-          let componentNode = nodes[0];
-          let propTypesNode = nodes[1];
-          let defaultPropsNode = nodes[2];
-          return actions.findPropTypes({
-            componentNode,
-            propTypesNode,
-            defaultPropsNode
-          }, options).then((propTypes) => {
-            if (propTypes.length === 0) {
-              throw new Error("Not find any props");
-            }
-            if (propTypesNode) {
-              if (propTypesNode.type === 'AssignmentExpression') {
-                // override old code style
-                options.codeStyle = 'default';
-              } else if (propTypesNode.type === 'ClassProperty') {
-                options.codeStyle = 'class';
-              }
-            } else if (componentNode.type === 'FunctionDeclaration' || componentNode.type === 'ArrowFunctionExpression') {
-              options.codeStyle = 'default';
-            }
-            let code = codeBuilder.buildPropTypes(propTypes, options);
-            if (propTypesNode) {
-              return replaceCode(filePath, propTypesNode.range, code);
-            } else {
-              let insertPosition;
-              if (options.codeStyle === 'class') {
-                if (componentNode.body) {
-                  insertPosition = componentNode.body.range[0] + 1;
-                  return insertCode(filePath, insertPosition, "\n  " + code + "\n");
-                }
-              } else {
-                insertPosition = componentNode.range[1];
-                return insertCode(filePath, insertPosition, "\n\n" + code);
-              }
-            }
-          });
-        }).then(result => {
-          if (options.autoImport !== 'disable') {
-            let { importNode, requireNode } = actions.findImportOrRequireModuleNode(ast, options);
-            let firstBody = ast.body[0];
-            let importCode = codeBuilder.buildImportCode(options);
-            if (!importNode && !requireNode && firstBody && importCode) {
-              let insertPosition = firstBody.range[0];
-              return insertCode(filePath, insertPosition, importCode + "\n");
-            }
-          }
-          return result;
-        }).then((result) => {
-          if (result) {
-            console.log('Generated Success!');
-          }
-        }).catch(error => {
-          console.error(error);
-        });
-      });
-    } catch (e) {
-      console.error(e);
-    }
+  .arguments('<filePath> [componentName]')
+  .action(function (filePath, componentName) {
+    filePath = path.join(process.cwd(), filePath || "")
+    parseAndGenerate(filePath, componentName)
   })
 
 program.command('config')
   .arguments('<filePath>')
-  .action(function(filePath){
+  .action(function (filePath) {
     let configPath = path.normalize(filePath);
     try {
       let json = fs.readFileSync(configPath, "utf-8") || {};
@@ -115,7 +38,127 @@ program.command('config')
     }
   })
 
+program.command('project')
+  .arguments('[dirPath]')
+  .action(function (dirPath) {
+    dirPath = path.join(process.cwd(), dirPath || "")
+    let files = getProjectJavascriptFiles(dirPath)
+    for (let i = 0; i < files.length; i++) {
+      parseAndGenerate(files[i])
+    }
+  })
+
 program.parse(process.argv)
+
+function getProjectJavascriptFiles(filePath) {
+  const files = []
+  const fileStat = fs.statSync(filePath)
+  if (fileStat.isDirectory()) {
+    const dirFiles = fs.readdirSync(filePath)
+    for (let i = 0; i < dirFiles.length; i++) {
+      files.push(...getProjectJavascriptFiles(path.join(filePath, dirFiles[i])))
+    }
+  } else {
+    if (filePath.match(/\.(js|jsx|ts|tsx)$/)) {
+      files.push(filePath)
+    }
+  }
+  return files;
+}
+
+function parseAndGenerate(filePath, componentName) {
+  let normalizePath = path.normalize(filePath);
+  let names = [];
+  if (componentName) {
+    names.push(componentName)
+  } else {
+    let data = fs.readFileSync(normalizePath, "utf-8");
+    if (data) {
+      let ast = astHelper.flowAst(data);
+      actions.findComponentNames(ast).forEach(item => {
+        names.push(item.name);
+      })
+    }
+  }
+  return Promise.reduce(names, function (total, name) {
+    return generatePropTypes(normalizePath, name).then(() => {
+      console.log(filePath + ' ' + name + ' Generated Success!');
+      return total + 1
+    }).catch(error => {
+      console.error(error)
+      return total
+    })
+  }, 0)
+}
+
+function generatePropTypes(filePath, componentName) {
+  let data = fs.readFileSync(filePath, "utf-8");
+  if (!data) {
+    return Promise.reject(new Error('can\'t resolve filePath: ' + filePath));
+  }
+  let ast = astHelper.flowAst(data);
+  let params = {
+    name: componentName
+  };
+  // merge config to options
+  let options = Object.assign({}, readConfig(), params);
+  return Promise.all([
+    actions.findComponentNode(ast, options),
+    actions.findPropTypesNode(ast, options),
+    actions.findPropTypesNode(ast, Object.assign({}, options, {
+      alias: 'defaultProps'
+    }))
+  ], options).then((nodes) => {
+    let componentNode = nodes[0];
+    let propTypesNode = nodes[1];
+    let defaultPropsNode = nodes[2];
+    return actions.findPropTypes({
+      componentNode,
+      propTypesNode,
+      defaultPropsNode
+    }, options).then((propTypes) => {
+      if (propTypes.length === 0) {
+        throw new Error("Not find any props");
+      }
+      if (propTypesNode) {
+        if (propTypesNode.type === 'AssignmentExpression') {
+          // override old code style
+          options.codeStyle = 'default';
+        } else if (propTypesNode.type === 'ClassProperty') {
+          options.codeStyle = 'class';
+        }
+      } else if (componentNode.type === 'FunctionDeclaration' || componentNode.type === 'ArrowFunctionExpression') {
+        options.codeStyle = 'default';
+      }
+      let code = codeBuilder.buildPropTypes(propTypes, options);
+      if (propTypesNode) {
+        return replaceCode(filePath, propTypesNode.range, code);
+      } else {
+        if (options.codeStyle === 'class') {
+          if (componentNode.body) {
+            return insertCode(filePath, componentNode.body.range[0] + 1, "\n  " + code + "\n");
+          }
+        } else {
+          let range = actions.findComponentParentRange(ast, componentName)
+          if (range) {
+            return insertCode(filePath, range[1], "\n\n" + code);
+          }
+        }
+      }
+    });
+  }).then(result => {
+    if (options.autoImport !== 'disable') {
+      let { importNode, requireNode } = actions.findImportOrRequireModuleNode(ast, options);
+      let firstBody = ast.body[0];
+      let importCode = codeBuilder.buildImportCode(options);
+      if (!importNode && !requireNode && firstBody && importCode) {
+        let insertPosition = firstBody.range[0];
+        return insertCode(filePath, insertPosition, importCode + "\n");
+      }
+    }
+    return result;
+  })
+}
 
 function readConfig() {
   try {