+++ src/actions/index.js
@@ -1,9 +1,12 @@
 const propTypes = require('./propTypes');
 const autoImport = require('./autoImport');
+const component = require('./component');
 
 module.exports = {
   findPropTypes: propTypes.findPropTypes,
   findPropTypesNode: propTypes.findPropTypesNode,
-  findComponentNode: propTypes.findComponentNode,
-  findImportOrRequireModuleNode: autoImport.findImportOrRequireModuleNode
+  findImportOrRequireModuleNode: autoImport.findImportOrRequireModuleNode,
+  findComponentNode: component.findComponentNode,
+  findComponentNames: component.findComponentNames,
+  findComponentParentRange: component.findComponentParentRange,
 };